style: squeeze text and center it on page
Reasoning
---------

The idea behind limiting the text's width is to make it easier for users
on large screens to read it.

Many professionals[1][2][3] recommend somewhere in the 60-80 words per
line, as such this commit adds a max width to text of 600px, which
should be roughly 60-70 words per line.

Technical implementation
------------------------

The technical implementation wraps every "text" content in a class, and
limits the width of that class's children. This is done as to not limit
the content of the whole page, and in particular the categories cards.

In addition the page's content is centered for aesthetic reasons.

[1]: https://www.w3.org/WAI/WCAG21/Understanding/visual-presentation.html#:~:text=For%20people%20with%20some,80%20characters%20or%20glyphs
[2]: http://web-accessibility.carnegiemuseums.org/design/font/#:~:text=Line%20length,longer%20length
[3]: https://www.uutilsynet.no/veiledning/tekst-og-struktur/226#:~:text=For%20best%20mulig%20lesbarhet,inkludert%20mellomrom